Add Auth

+++ store-api/routes/api.php
@@ -13,12 +13,22 @@ use Illuminate\Http\Request;
 |
 */
 
-Route::get('products/{title?}/{likes?}', 'ProductController@index');
-Route::get('products/{product}', 'ProductController@show');
-Route::post('products', 'ProductController@store');
-Route::put('products/{product}', 'ProductController@update');
-Route::delete('products/{id}', 'ProductController@delete');
+Route::get('products/{name?}', 'ProductController@index');
+Route::get('products/like/{id}', 'ProductController@like');
+Route::post('products', 'ProductController@store')->middleware('auth:api');
+Route::put('products/{product}', 'ProductController@update')->middleware('auth:api');
+Route::delete('products/{id}', 'ProductController@delete')->middleware('auth:api');
 
-Route::middleware('auth:api')->get('/user', function (Request $request) {
-    return $request->user();
+Route::group([
+    'prefix' => 'auth'
+], function () {
+    Route::post('login', 'AuthController@login');
+    Route::post('signup', 'AuthController@signup');
+
+    Route::group([
+        'middleware' => 'auth:api'
+    ], function () {
+        Route::get('logout', 'AuthController@logout');
+        Route::get('user', 'AuthController@user');
+    });
 });
